renamed caching.episode --> caching.short_term
